strands/quickstart_demo_agent/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
python-dotenv>=1.0.0 
boto3